% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalization}
\usage{
normalize(x, method = c("norm", "gauss", "minmax"))
}
\arguments{
\item{x}{Matrix with the performances of each alternative at each criterion}

\item{method}{Character string. Normalization method. Either "norm", "gauss" or "minmax".}
}
\description{
Normalization of the performance matrix.
}
\examples{

x <- matrix(1:16, nrow = 4)
normalize(x)

}
\author{
\strong{Rafael Benítez} (\email{rafael.suarez@uv.es}).
\emph{Department of Business Mathematics}

\strong{Vicente Liern} (\email{vicente.liern@uv.es}).
\emph{Department of Business Mathematics}

University of Valencia (Spain)
}
