% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPI}
\alias{SPI}
\title{Data: P. Annoni, L. Dijkstra and T. E. Hellman (2016).}
\format{Data frame (tibble) with 272 rows and 13 columns. The variables are:
\describe{
  \item{Region Code}{NUTS 1, 2 or 3 code for the European region to be evaluated.}
  \item{Columns 2 - 5}{Components of the "Basic Human Needs" dimension.}
  \item{Columns 6 - 9}{Components of the "Foundations of Wellbeing" dimension.}
  \item{Columns 10 - 13}{Components of the "Opportunity" dimension.}
}}
\source{
P. Annoni, L. Dijkstra and T. E. Hellman (2016). "The EU Regional SPI: a Measure of Social Progress in the EU Regions", Methodological Paper. \code{url}: http://ec.europa.eu/regional_policy/sources/information/maps/methodological_note_eu_spi_2016.pdf. Accesed on 2020-03-08
}
\usage{
data("SPI")
}
\description{
Components of the three dimensions used to obtain the Social Progress Index for European Regions in 2016.
}
\author{
\strong{Vicente Liern Carrion} (\email{vicente.liern@uv.es}).
\emph{Department of Business Mathematics}

\strong{Rafael Benitez} (\email{rafael.suarez@uv.es}).
\emph{Department of Business Mathematics}

University of Valencia (Spain)
}
\keyword{datasets}
